// src/app/state/detectors/finger-types.enum.ts
export enum FingerTypes {
  Wrist = 'wrist',
  ThumbCMC = 'thumb_cmc',
  ThumbMCP = 'thumb_mcp',
  ThumbIP = 'thumb_ip',
  ThumbTip = 'thumb_tip',
  IndexFingerMCP = 'index_finger_mcp',
  IndexFingerPIP = 'index_finger_pip',
  IndexFingerDIP = 'index_finger_dip',
  IndexFingerTip = 'index_finger_tip',
  MiddleFingerMCP = 'middle_finger_mcp',
  MiddleFingerPIP = 'middle_finger_pip',
  MiddleFingerDIP = 'middle_finger_dip',
  MiddleFingerTip = 'middle_finger_tip',
  RingFingerMCP = 'ring_finger_mcp',
  RingFingerPIP = 'ring_finger_pip',
  RingFingerDIP = 'ring_finger_dip',
  RingFingerTip = 'ring_finger_tip',
  PinkyFingerMCP = 'pinky_finger_mcp',
  PinkyFingerPIP = 'pinky_finger_pip',
  PinkyFingerDIP = 'pinky_finger_dip',
  PinkyFingerTip = 'pinky_finger_tip',
}